Perubahan halaman Ingredient

+++ database/migrations/2026_04_11_000001_create_ingredients_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->enum('unit', ['gram', 'kg', 'ml', 'liter', 'pcs', 'sachet', 'sdm', 'sdt']);
-            $table->decimal('low_stock_threshold', 12, 2)->default(0);
+            $table->integer('low_stock_threshold')->default(0);
             $table->boolean('is_active')->default(true);
             $table->timestamps();
             $table->softDeletes();
